feat: masque la date de certains assistants
