+++ site/source/components/DateChip.tsx
@@ -1,0 +1,32 @@
+import { useTranslation } from 'react-i18next'
+
+import { Chip } from '@/design-system'
+import { Emoji } from '@/design-system/emoji'
+import { useCurrentSimulatorData } from '@/hooks/useCurrentSimulatorData'
+import useDate from '@/hooks/useDate'
+
+export default function DateChip() {
+	const { currentSimulatorData } = useCurrentSimulatorData()
+	const showDate = !currentSimulatorData?.hideDate
+
+	const engineDate = useDate()
+	const date = engineDate?.toString().slice(-7)
+
+	const { t } = useTranslation()
+
+	return (
+		showDate &&
+		date && (
+			<Chip
+				type="secondary"
+				icon={<Emoji emoji="📆" />}
+				title={t(
+					'pages.simulateurs.date',
+					'Date de la réglemenation utilisée pour les calculs'
+				)}
+			>
+				{date}
+			</Chip>
+		)
+	)
+}
