+++ site/source/pages/simulateurs/impot-societe/config.tsx
@@ -29,8 +29,10 @@ export function impôtSociétéConfig({ t, sitePaths }: SimulatorsDataParams) {
 			'pages.simulateurs.impot-société.title',
 			'Simulateur d’impôt sur les sociétés'
 		),
+		date: false,
 		nextSteps: ['salarié', 'comparaison-statuts'],
 		path: sitePaths.simulateurs.is,
+		hideDate: true,
 		component: ISSimulation,
 		seoExplanations: SeoExplanations,
 		simulation: ISSimulationConfig,
